+++ lib/src/material/amber/amber_accents.dart
@@ -7,7 +7,7 @@ import 'amber_accent.dart';
 /// See also
 /// - [amber accent](https://api.flutter.dev/flutter/material/Colors/amberAccent-constant.html)
 class AmberAccents extends SwatchBase {
-  /// Four shades of amber accent; the greater the index, the darker the color.
+  /// Four shades of amber accent; the higher the index, the darker the color.
   ///
   /// There are 4 valid indexes
   /// - 0, light ≡ 100