+++ lib/src/material/blue/blues.dart
@@ -7,7 +7,7 @@ import 'blue.dart';
 /// See also
 /// - [blue](https://api.flutter.dev/flutter/material/Colors/blue-constant.html)
 class Blues extends SwatchBase {
-  /// Ten shades of blue; the greater the index, the darker the color.
+  /// Ten shades of blue; the higher the index, the darker the color.
   ///
   /// There are 10 valid indexes
   /// - 0, ultra light ≡ 50