+++ lib/src/material/light_blue/light_blue_accents.dart
@@ -13,7 +13,7 @@ import 'light_blue_accent.dart';
 /// See also
 /// - [light blue accent](https://api.flutter.dev/flutter/material/Colors/lightBlueAccent-constant.html)
 class LightBlueAccents extends SwatchBase {
-  /// Four shades of light blue accent; the greater the index, the darker the color.
+  /// Four shades of light blue accent; the higher the index, the darker the color.
   ///
   /// There are 4 valid indexes
   /// - 0, light ≡ 100