+++ lib/src/material/light_green/light_green_accents.dart
@@ -13,7 +13,7 @@ import 'light_green_accent.dart';
 /// See also
 /// - [light green accent](https://api.flutter.dev/flutter/material/Colors/lightGreenAccent-constant.html)
 class LightGreenAccents extends SwatchBase {
-  /// Four shades of light green accent; the greater the index, the darker the color.
+  /// Four shades of light green accent; the higher the index, the darker the color.
   ///
   /// There are 4 valid indexes
   /// - 0, light ≡ 100