+++ lib/src/material/light_green/light_greens.dart
@@ -7,7 +7,7 @@ import 'light_green.dart';
 /// See also
 /// - [lightGreen](https://api.flutter.dev/flutter/material/Colors/lightGreen-constant.html)
 class LightGreens extends SwatchBase {
-  /// Ten shades of light green; the greater the index, the darker the color.
+  /// Ten shades of light green; the higher the index, the darker the color.
   ///
   /// There are 10 valid indexes
   /// - 0, ultra light ≡ 50