+++ lib/src/material/teal/teal_accents.dart
@@ -7,7 +7,7 @@ import 'teal_accent.dart';
 /// See also
 /// - [teal accent](https://api.flutter.dev/flutter/material/Colors/tealAccent-constant.html)
 class TealAccents extends SwatchBase {
-  /// Four shades of teal accent; the greater the index, the darker the color.
+  /// Four shades of teal accent; the higher the index, the darker the color.
   ///
   /// There are 4 valid indexes
   /// - 0, light ≡ 100